requirements.txt
numpy
openai-whisper
ffmpeg
ffprobe
